Initial makefile
I copied over the MRI makefile since it builds both unit tests to run
on the host and final code to run on the Cortex-M prcocessors.

The ARM target based portion of the makefile is mostly commented out at
this point and I will add it in shortly.  So far I have been
concentrating on the unit tests only.