Make sure that stack doesn't overflow
CrashCatcher uses its own stack when creating a crash dump.  There is
a risk that this dedicated stack is too small (currently set to 1k).
This commit adds a simple overflow check which sets the bottom word of
the stack to 0xACCE55ED at the beginning of the core CrashCatcher loop
and then makes sure that it still has the same value at the end of each
loop iteration.

If an overflow is detected, an extra 4 bytes of ACCE55ED are appended
to the crash dump.