Make stack overflow message easier to read
Previously I was dumping 0xACCE55ED as a word but due to little endian
the bytes in the dump aren't as obvious.  I switched to using a byte
array for this value instead so that it would be more obvious when
looking at a hex representation of a dump whether a stack overflow
has been detected.