Adding mbed1768 test application
Adding CrashingApp as a sample application that links in the
CrashCatcher module.  It lets the user pick one of the following
scenarios before issuing a bad read to cause a crash:
* Using MSP that is 8-byte aligned at time of crash.
* Using MSP that is NOT 8-byte aligned at time of crash.
* Using PSP that is 8-byte aligned at time of crash.

I currently just have it building and running on the mbed1768 device.
I will add the mbed11u24 device in the future.

NOTE: The makefile for this sample expects GCC4MBED to be installed
      in a directory which is a sibling of this CrashCatcher
      repository.