Add LocalFileSystem support
The original mbed devices such as the mbed-LPC1768 and mbed-LPC11U24
contain a 2MB FLASH drive which is accessible to code running on the
Cortex-M processor via the LocalFileSystem APIs.  The contents placed
in this file system can later be accessed through the USB MSD device
mounted when the mbed is plugged into the PC's USB port.

The libCrashCatcher_LocalFileSystem_*.a libraries can be used to create
a CRASH.DMP file on this LocalFileSystem when a crash occurs.  The user
can later extract it by plugging the mbed device into a PC.

I also create a LocalFileSystemTest sample that links with this
library and tested it on my mbed-LPC1768.