Get code building on Windows
I switched from using INTPTR_MAX and INT64_MAX for determining if code
is going to run on a 64-bit machine and just switched to sizeof()
operator on stdint types instead.  This was easier to get working in
C++ code across compilers/platforms.

makedir calls in makefile should always have ended with $(QUIET)