Update to version 3, adds MSP & PSP
The crash dumps now include the MSP and PSP register contents at the
time of exception entry. This is useful for debugging crashes that
occur in a handler while running an RTOS where the interrupted user
code callstack can only be found via the PSP register.

This commit also includes:
* Now includes gcc4mbed as submodule so that the samples can be built
  with the same branch of this project as I used previously.
* I increased CRASH_CATCHER_STACK_WORD_COUNT to allocate more space to
  the stack used while executing CrashCatcher's fault handler.
* The major version in the dump header is updated to 3, indicating that
  the integer registers will now include MSP and PSP registers. These
  registers will follow PSR and precede exceptionPSR.
* Updated to pull the latest CppUTest bug fixes.
* I updated README to document the changes in the crash dump format
  introduced by this version update.