Allow CRASH_CATCHER_STACK_WORD_COUNT override
Thanks to Fredrik for pointing out that it would be advantageous to be
able to set this constant without modifying the source code itself.
This macro defines how large of a stack to reserve for the code which
runs when CrashCatcher is dumping a crash log.

I also added a "CrashCatcher Stack" section to the README to give an
overview of the stack used by CrashCatcher and how to increase it.