Disable BKPT check by default
Catching and dumping faults is higher priority than detecting and
allowing restarts from hardcoded BKPTs so I have disabled this feature
by default. This means that the code to check the instruction at a
potentially bad PC (that lead to the fault) is disabled and will no
longer lead to a double fault.

It can be re-enabled by setting CRASH_CATCHER_ISBKPT_SUPPORT to 1 in
CrashCatcherPriv.h or on the compiler command line.